More refactor to remove conv2dVec4 program
